Move two text fixture to DynamoCoreWpfTests namespace
